imu: add debug-interface to DMU driver
This adds a simple hook for reading DMU data from a file instead of
opening up a device. Messages are constructed and sent to publisher.

Currently batching is set to 20 and _no_ effort is spent trying to
detect if we miss a sampleset from the device, this is still on the
TOO-list.