imu_helper: do correct byteswap for floats
Without a union-wrapper, htobe32 will cast to uint32_t *before*
dereferencing the pointer causing integer truncation of the value.

This means that the function either needs to be passed a uint32_t
pointer, or it must take a float and copy it into a union that also
expses a uint32_t value, which can be byteswapped correctly.

The latter is a few extra cycles, but it provides code that is easier
to read and simpler asm.