IMU: make sure device exists before starting IMU
Default value for IMU is /dev/ttyUSB0, but no checks are made upon
startup if this is actually a valid path to a valid device. The same
goes for the debug_file argument.

Let the IMU instance ascertain validity and test this before starting
the server to avoid errors once the device is activated and i3ds_imu
attempts to open the device/file.